Revert "feat(view): unify segmented control styling"
