+++ OffshoreBudgeting/Views/BudgetDetailsView.swift
@@ -344,7 +344,12 @@ private extension BudgetDetailsView {
                 .pickerStyle(.segmented)
                 .equalWidthSegments()
                 .frame(maxWidth: .infinity)
-                .liquidAwareSegmentedControlStyle(accentColor: themeManager.selectedTheme.glassPalette.accent)
+#if os(macOS)
+                .controlSize(.large)
+
+                .tint(themeManager.selectedTheme.glassPalette.accent)
+
+#endif
             }
             .padding(.horizontal, DS.Spacing.l)
             .ub_onChange(of: vm.selectedSegment) { newValue in
@@ -729,7 +734,12 @@ private struct FilterBar: View {
             }
             .pickerStyle(.segmented)
             .equalWidthSegments()
-            .liquidAwareSegmentedControlStyle(accentColor: themeManager.selectedTheme.glassPalette.accent)
+#if os(macOS)
+            .controlSize(.large)
+
+            .tint(themeManager.selectedTheme.glassPalette.accent)
+
+#endif
             .frame(maxWidth: .infinity)
         }
         .frame(maxWidth: .infinity)
