+++ OffshoreBudgeting/Views/HomeView.swift
@@ -683,7 +683,7 @@ struct HomeView: View {
                     .pickerStyle(.segmented)
                     .equalWidthSegments()
                     .frame(maxWidth: .infinity)
-                    .liquidAwareSegmentedControlStyle(accentColor: themeManager.selectedTheme.glassPalette.accent)
+                    .modifier(homeSegmentedControlStyling)
                 }
 
                 // Filter bar (sort options)
@@ -698,7 +698,7 @@ struct HomeView: View {
                     .pickerStyle(.segmented)
                     .equalWidthSegments()
                     .frame(maxWidth: .infinity)
-                    .liquidAwareSegmentedControlStyle(accentColor: themeManager.selectedTheme.glassPalette.accent)
+                    .modifier(homeSegmentedControlStyling)
                 }
 
                 // Always-offer Add button when no budget exists so users can
@@ -728,6 +728,13 @@ struct HomeView: View {
         }
         .ub_ignoreSafeArea(edges: .bottom)
         .ub_hideScrollIndicators()
+    }
+
+    private var homeSegmentedControlStyling: HomeSegmentedControlModifier {
+        HomeSegmentedControlModifier(
+            capabilities: capabilities,
+            accentColor: themeManager.selectedTheme.glassPalette.accent
+        )
     }
 
     private var headerSectionSpacing: CGFloat {
@@ -1310,6 +1317,25 @@ private struct HomeHeaderMinWidthModifier: ViewModifier {
 
     private var minimumWidth: CGFloat {
         RootHeaderActionMetrics.minimumGlassWidth(for: capabilities)
+    }
+}
+
+private struct HomeSegmentedControlModifier: ViewModifier {
+    let capabilities: PlatformCapabilities
+    let accentColor: Color
+
+    func body(content: Content) -> some View {
+#if os(macOS)
+        if capabilities.supportsOS26Translucency {
+            content
+        } else {
+            content
+                .controlSize(.large)
+                .tint(accentColor)
+        }
+#else
+        content
+#endif
     }
 }
